Reworked PostRestController and PostService to add authentication pro…

+++ src/main/java/com/example/carforum/services/PostServiceImpl.java
@@ -2,33 +2,35 @@ package com.example.carforum.services;
 
 import com.example.carforum.exceptions.EntityNotFoundException;
 import com.example.carforum.models.Post;
+import com.example.carforum.models.User;
 import com.example.carforum.repositories.PostRepository;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
+import org.springframework.web.server.ResponseStatusException;
 
 import java.util.List;
 
 @Service
-public class PostServiceImpl implements PostService{
+public class PostServiceImpl implements PostService {
 
 
     private static final String NOT_FOUND_MESSAGE = "%s with %d id was not found!";
+    public static final String OWNER_ERROR_MESSAGE = "You are not authorized to edit/delete posts created by other users.";
     private final PostRepository repository;
 
     @Autowired
-    public PostServiceImpl(PostRepository repository){
+    public PostServiceImpl(PostRepository repository) {
         this.repository = repository;
     }
 
     @Override
     public Post getById(int id) {
-
         Post post = repository.getById(id);
 
-        if (post == null){
-
-            throw new EntityNotFoundException(String.format(NOT_FOUND_MESSAGE,  "Post", id));
-        }else {
+        if (post == null) {
+            throw new EntityNotFoundException(String.format(NOT_FOUND_MESSAGE, "Post", id));
+        } else {
             return post;
         }
     }
@@ -45,22 +47,23 @@ public class PostServiceImpl implements PostService{
 
     @Override
     public void create(Post post) {
-
         repository.create(post);
-
     }
 
     @Override
-    public void update(Post post) {
-
+    public void update(Post post, User user) {
+        if (post.getUser().getId() != user.getId()) {
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN, OWNER_ERROR_MESSAGE);
+        }
         repository.update(post);
     }
 
     @Override
-    public void deleteById(int id) {
-
+    public void deleteById(int id, User user) {
         Post post = getById(id);
-
+        if (post.getUser().getId() != user.getId()) {
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN, OWNER_ERROR_MESSAGE);
+        }
         repository.delete(post);
     }
 }
